🚀 feat(ProcessEPGFileRequest.cs): remove unused using statements The using statements for StreamMasterApplication.Services and StreamMasterDomain.Entities were removed as they were not being used in the file.
🚀 feat(AddProgrammesFromSDRequest.cs): add new file
A new file was added to the project to handle the request to add programmes from Schedules Direct. The file contains a request record, a validator, and a handler. The handler uses the SchedulesDirect NuGet package to retrieve programme information from the Schedules Direct API.

🚀 feat(IProgrammeChannelInterfaces.cs): add AddProgrammesFromSDRequest to IProgrammeChannelTasks
The IProgrammeChannelTasks interface was updated to include a new method, AddProgrammesFromSDRequest, which will be used to add programmes from Schedules Direct.

🚀 feat(StreamMasterApplication.csproj): add reference to StreamMaster.SchedulesDirect project
The StreamMaster.SchedulesDirect project was added as a reference to the StreamMasterApplication project.

🚀 feat(SettingDto.cs): update version number
The version number in the SettingDto class was updated to 0.2.4.

🚀 feat(SMQueCommand.cs): add AddProgrammes