feat: Add option to format EPG output
Added a new boolean field "PrettyEPG" to the UpdateSettingRequest class the user to enable or disable the formatting of the EPG output. The default value is set to false.

Modified the GetStreamGroupEPGHandler class to use value of the PrettyEPG field whenizing the XMLTV data. If the PrettyEPG field is true, the XML output will be indented. If it is false, the XML output will not be indented.

Modified the FilesEPGM3USettings component to display a checkbox for the PrettyEPG field.

Modified the iptvApi.ts file to the "prettyEPG" field in the BaseSettings type and in the UpdateSettingRequest type.

Modified the MessagesEn.tsx file to include the translation for the "prettyEPG" field.