fix: Check if the memory read is equal to the video buffer size
This commit fixes a bug in the StreamHandler class where it was not checking if the memory read from the circular ring buffer was equal to the video buffer size. If the memory read is not equal to the video buffer size, the method will now return instead of continuing execution. This bug fix will trigger a release bumping a PATCH version.