feat: Add support for database settings
This commit adds support for configuring the database settings in the application. It introduces a new `DBSettings` class that contains properties for the database host, name, user, and password. The ` class now includes an instance of `DBSettings` called `DB` to store the database configuration.

The `LogDbContext` and `RepositoryContext` classes have been updated to use the database settings from the `Setting` class when creating the database connection string. The connection string is now constructed dynamically based on the values the `DBSettings` instance.

This new feature allow users to easily configure the database connection for the application.