fix: Handle case when toRead is less than 0 in StreamHandler
Previously, did not handle the case when `toRead` was less than 0 in the `StreamHandler` class. This commit fixes that by adding a condition to check if `toRead` is less than and logging an error if it is. If `toRead` is greater than or equal to  the code proceeds as before. This fix ensures that the program handles this edge case correctly and improves the overall reliability of the `StreamHandler` class.